+++ build_runner/lib/src/generate/build_definition.dart
@@ -7,6 +7,7 @@ import 'dart:convert';
 import 'dart:io';
 
 import 'package:build/build.dart';
+import 'package:build_config/build_config.dart';
 import 'package:glob/glob.dart';
 import 'package:logging/logging.dart';
 import 'package:meta/meta.dart';
@@ -57,18 +58,21 @@ class BuildDefinition {
       this.enableLowResourcesMode,
       this.onDelete);
 
-  static Future<BuildDefinition> prepareWorkspace(
-          BuildOptions options, List<BuildAction> buildActions,
+  static Future<BuildDefinition> prepareWorkspace(BuildOptions options,
+          List<BuildAction> buildActions, BuildConfig rootPackageConfig,
           {void onDelete(AssetId id)}) =>
-      new _Loader(options, buildActions, onDelete).prepareWorkspace();
+      new _Loader(options, buildActions, rootPackageConfig, onDelete)
+          .prepareWorkspace();
 }
 
 class _Loader {
   final List<BuildAction> _buildActions;
   final BuildOptions _options;
   final OnDelete _onDelete;
+  final BuildConfig _rootPackageConfig;
 
-  _Loader(this._options, this._buildActions, this._onDelete);
+  _Loader(this._options, this._buildActions, this._rootPackageConfig,
+      this._onDelete);
 
   Future<BuildDefinition> prepareWorkspace() async {
     _checkBuildActions();
@@ -329,9 +333,8 @@ class _Loader {
       yield* _options.reader.findAssets(new Glob(glob), package: package.name);
     }
     if (!package.isRoot) return;
-    for (final action
-        in _buildActions.where((a) => a.package == package.name)) {
-      for (final glob in action.generateFor?.include ?? const []) {
+    for (final target in _rootPackageConfig?.buildTargets?.values ?? const []) {
+      for (final glob in target.sources.include ?? const []) {
         yield* _options.reader
             .findAssets(new Glob(glob), package: package.name);
       }